Move all constraints to the DraftFour namespace.
