+++ src/Constraints/DraftFour/ExclusiveMinimum.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace League\JsonGuard\Constraints;
+namespace League\JsonGuard\Constraints\DraftFour;
 
 use League\JsonGuard\Assert;
-use function League\JsonGuard\error;
+use League\JsonGuard\Constraint;
 use League\JsonGuard\Validator;
+use function League\JsonGuard\error;
 
 class ExclusiveMinimum implements Constraint
 {
