+++ src/Constraints/DraftFour/MinLength.php
@@ -1,15 +1,16 @@
 <?php
 
-namespace League\JsonGuard\Constraints;
+namespace League\JsonGuard\Constraints\DraftFour;
 
 use League\JsonGuard;
 use League\JsonGuard\Assert;
-use function League\JsonGuard\error;
+use League\JsonGuard\Constraint;
 use League\JsonGuard\Validator;
+use function League\JsonGuard\error;
 
-class MaxLength implements Constraint
+class MinLength implements Constraint
 {
-    const KEYWORD = 'maxLength';
+    const KEYWORD = 'minLength';
 
     /**
      * @var string
@@ -24,7 +25,6 @@ class MaxLength implements Constraint
         $this->charset = $charset;
     }
 
-
     /**
      * {@inheritdoc}
      */
@@ -33,10 +33,10 @@ class MaxLength implements Constraint
         Assert::type($parameter, 'number', self::KEYWORD, $validator->getSchemaPath());
         Assert::nonNegative($parameter, self::KEYWORD, $validator->getSchemaPath());
 
-        if (!is_string($value) || JsonGuard\strlen($value, $this->charset) <= $parameter) {
+        if (!is_string($value) || JsonGuard\strlen($value, $this->charset) >= $parameter) {
             return null;
         }
 
-        return error('The string must be less than {parameter} characters long.', $validator);
+        return error('The string must be at least {parameter} characters long.', $validator);
     }
 }
