+++ tests/ValidatorTest.php
@@ -3,17 +3,17 @@
 namespace League\JsonGuard\Test;
 
 use League\JsonGuard;
-use League\JsonGuard\Constraints\Constraint;
-use function League\JsonGuard\error;
-use League\JsonReference\Dereferencer;
+use League\JsonGuard\Constraint;
+use League\JsonGuard\Constraints\DraftFour\Format\FormatExtension;
 use League\JsonGuard\Exceptions\InvalidSchemaException;
 use League\JsonGuard\Exceptions\MaximumDepthExceededException;
-use League\JsonGuard\FormatExtension;
+use League\JsonGuard\RuleSets\DraftFour;
+use League\JsonGuard\Validator;
+use League\JsonReference\Dereferencer;
 use League\JsonReference\Loaders\ArrayLoader;
 use League\JsonReference\Loaders\ChainableLoader;
 use League\JsonReference\Loaders\CurlWebLoader;
-use League\JsonGuard\Validator;
-use League\JsonGuard\RuleSets\DraftFour;
+use function League\JsonGuard\error;
 
 class ValidatorTest extends \PHPUnit_Framework_TestCase
 {
@@ -132,7 +132,7 @@ class ValidatorTest extends \PHPUnit_Framework_TestCase
         $this->assertCount(2, $errors);
 
         $this->assertSame(
-            JsonGuard\Constraints\Type::KEYWORD,
+            JsonGuard\Constraints\DraftFour\Type::KEYWORD,
             $errors[0]->getKeyword(),
             'The error should include the keyword'
         );
@@ -209,7 +209,7 @@ class ValidatorTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($v->fails());
         $error = $v->errors()[0];
         $this->assertSame('/foo/foo/foo/foo/foo/foo/foo/foo/foo', $error->getDataPath());
-        $this->assertSame(JsonGuard\Constraints\AdditionalProperties::KEYWORD, $error->getKeyword());
+        $this->assertSame(JsonGuard\Constraints\DraftFour\AdditionalProperties::KEYWORD, $error->getKeyword());
     }
 
     function test_it__throws_max_depth_exception_when_max_depth_is_exceeded()
@@ -256,7 +256,7 @@ class ValidatorTest extends \PHPUnit_Framework_TestCase
         $v->getRuleSet()->get('format')->addExtension('hello', new HelloFormatStub());
 
         $this->assertTrue($v->fails());
-        $this->assertSame(JsonGuard\Constraints\Format::KEYWORD, $v->errors()[0]->getKeyword());
+        $this->assertSame(JsonGuard\Constraints\DraftFour\Format::KEYWORD, $v->errors()[0]->getKeyword());
     }
 
     function test_custom_format_works_when_nested()
